Fix search crashing on pharmacies that have ratings
The rating aggregate used an uncast AVG(), which Postgres returns as
numeric; Prisma hands that back as a Decimal and JSON-encodes it as a
string. RatingStars then called .toFixed() on a string, throwing
during render and blanking the whole page with "This page couldn't
load" as soon as a result carried a rating.

Casts the average to float8 so it arrives as a number, and hardens
RatingStars to coerce and fall back to "unrated" rather than throw —
a rating value should never be able to take down search results.

+++ src/lib/geo.ts
@@ -66,7 +66,9 @@ export async function findPharmaciesWithDrug(opts: {
       SELECT
         "pharmacyId",
         COUNT(*) AS "ratingCount",
-        AVG(("availability" + "service" + "pricing" + "honesty") / 4.0) AS "ratingAvg"
+        -- ::float8 matters: an uncast AVG() is numeric, which Prisma hands
+        -- back as a Decimal and JSON-encodes as a string, not a number.
+        AVG(("availability" + "service" + "pricing" + "honesty") / 4.0)::float8 AS "ratingAvg"
       FROM "PharmacyRating"
       GROUP BY "pharmacyId"
     ) r ON r."pharmacyId" = p."id"
